requirements.txt
streamlit>=1.30.0
langchain>=0.3.0
langchain-classic>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
pypdf>=4.0.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
